tcap_v3/requirements.txt
# TCAP v3 Requirements
# Core dependencies for automated cryptocurrency trading

# HTTP and API
aiohttp>=3.8.0
requests>=2.25.0

# Data Analysis
pandas>=1.3.0
numpy>=1.21.0

# Technical Analysis
ta>=0.7.0  # Technical Analysis library

# Database
sqlite3  # Built into Python

# Logging and Configuration
python-dotenv>=0.19.0

# Async Support
asyncio  # Built into Python 3.7+

# Cryptography for API signing
cryptography>=3.4.8

# Testing (optional)
pytest>=6.2.0
pytest-asyncio>=0.15.0

# Development tools (optional)
black>=21.0.0
flake8>=3.9.0
